+++ src/training/unicharset/unicharset_training_utils.cpp
@@ -154,7 +154,7 @@ void SetScriptProperties(const std::string &script_dir, UNICHARSET *unicharset) 
       tprintf("Failed to load script unicharset from:%s\n", filename.c_str());
     }
   }
-  for (int c = SPECIAL_UNICHAR_CODES_COUNT; c < unicharset->size(); ++c) {
+  for (size_t c = SPECIAL_UNICHAR_CODES_COUNT; c < unicharset->size(); ++c) {
     if (unicharset->PropertiesIncomplete(c)) {
       tprintf("Warning: properties incomplete for index %d = %s\n", c,
               unicharset->id_to_unichar(c));
